refactor: simplify `op-proposer` logic (#79)
* wip

* add

* fix l1Client

* add

* add comments

* update

* update

* fix: precompile override

* add value

* docs

---------
