+++ .env.server.example
@@ -4,6 +4,11 @@ L1_BEACON_RPC=
 L2_RPC=
 L2_NODE_RPC=
 
+L1_ETH_RPC=
+BEACON_RPC=
+ROLLUP_RPC=
+
+
 # op-proposer configuration
 POLL_INTERVAL=
 L2OO_ADDRESS=
