+++ client-programs/range/src/main.rs
@@ -72,8 +72,6 @@ fn main() {
                 let boot = Arc::new(BootInfo::load(oracle.as_ref()).await.unwrap());
             }
         }
-        // Note: On some blocks, key not found in cache errors occur due to the precompiles. For
-        // recent blocks this isn't an issue, but we should look into this more in the future.
         let precompile_overrides = ZKVMPrecompileOverride::default();
 
         let l1_provider = OracleL1ChainProvider::new(boot.clone(), oracle.clone());
@@ -127,6 +125,11 @@ fn main() {
             let new_block_number = new_block_header.number;
             assert_eq!(new_block_number, payload.parent.block_info.number + 1);
 
+            // Increment last_block_num and check if we have reached the claim block.
+            if new_block_number == boot.l2_claim_block {
+                break 'step;
+            }
+
             // Generate the Payload Envelope, which can be used to derive cached data.
             let l2_payload_envelope: L2ExecutionPayloadEnvelope = OpBlock {
                 header: new_block_header.clone(),
@@ -148,11 +151,6 @@ fn main() {
             l2_block_info = l2_provider
                 .update_cache(new_block_header, l2_payload_envelope, &boot.rollup_config)
                 .unwrap();
-
-            // Increment last_block_num and check if we have reached the claim block.
-            if new_block_number == boot.l2_claim_block {
-                break 'step;
-            }
 
             // Update data for the next iteration.
             driver.update_safe_head(
