+++ op-proposer-go/cmd/main.go
@@ -5,8 +5,10 @@ import (
 	"log"
 	"os"
 
+	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/ethereum-optimism/optimism/op-service/dial"
 	"github.com/joho/godotenv"
-	"github.com/succinctlabs/op-succinct-go/server/utils"
+	"github.com/succinctlabs/op-succinct-go/proposer/utils"
 	"github.com/urfave/cli/v2"
 )
 
@@ -63,6 +65,20 @@ func main() {
 				log.Fatal(err)
 			}
 
+			l1BeaconClient, err := utils.SetupBeacon(cliCtx.String("l1.beacon"))
+			if err != nil {
+				log.Fatal(err)
+			}
+
+			l1Client, err := ethclient.Dial(cliCtx.String("l1"))
+			if err != nil {
+				log.Fatal(err)
+			}
+			rollupClient, err := dial.DialRollupClientWithTimeout(cliCtx.Context, dial.DefaultDialTimeout, nil, cliCtx.String("l2.node"))
+			if err != nil {
+				log.Fatal(err)
+			}
+
 			config := utils.BatchDecoderConfig{
 				L2GenesisTime:     rollupCfg.Genesis.L2Time,
 				L2GenesisBlock:    rollupCfg.Genesis.L2.Number,
@@ -71,14 +87,14 @@ func main() {
 				L2StartBlock:      cliCtx.Uint64("start"),
 				L2EndBlock:        cliCtx.Uint64("end"),
 				L2ChainID:         rollupCfg.L2ChainID,
-				L2Node:            cliCtx.String("l2.node"),
-				L1RPC:             cliCtx.String("l1"),
-				L1Beacon:          cliCtx.String("l1.beacon"),
+				L2Node:            rollupClient,
+				L1RPC:             *l1Client,
+				L1Beacon:          l1BeaconClient,
 				BatchSender:       rollupCfg.Genesis.SystemConfig.BatcherAddr,
 				DataDir:           fmt.Sprintf("/tmp/batch_decoder/%d/transactions_cache", rollupCfg.L2ChainID),
 			}
 
-			ranges, err := utils.GetAllSpanBatchesInBlockRange(config)
+			ranges, err := utils.GetAllSpanBatchesInL2BlockRange(config)
 			if err != nil {
 				log.Fatal(err)
 			}
