+++ op-proposer-go/proposer/driver.go
@@ -151,6 +151,7 @@ func newL2OOSubmitter(ctx context.Context, cancel context.CancelFunc, setup Driv
 	}, nil
 }
 
+// Create a new submitter for the DisputeGameFactory. Note: This is unused in OP-Succinct.
 func newDGFSubmitter(ctx context.Context, cancel context.CancelFunc, setup DriverSetup) (*L2OutputSubmitter, error) {
 	dgfCaller, err := bindings.NewDisputeGameFactoryCaller(*setup.Cfg.DisputeGameFactoryAddr, setup.L1Client)
 	if err != nil {
