+++ op-proposer-go/server/main.go
@@ -8,7 +8,9 @@ import (
 	"net/http"
 	"sort"
 
-	"github.com/succinctlabs/op-succinct-go/server/utils"
+	"github.com/ethereum-optimism/optimism/op-service/dial"
+	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/succinctlabs/op-succinct-go/proposer/utils"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/gorilla/mux"
@@ -46,18 +48,39 @@ func handleSpanBatchRanges(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	l1BeaconClient, err := utils.SetupBeacon(req.L1Beacon)
+	if err != nil {
+		fmt.Printf("Error setting up beacon: %v\n", err)
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	l1Client, err := ethclient.Dial(req.L1RPC)
+	if err != nil {
+		fmt.Printf("Error creating L1 client: %v\n", err)
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	l2Node, err := dial.DialRollupClientWithTimeout(r.Context(), dial.DefaultDialTimeout, nil, req.L2Node)
+	if err != nil {
+		fmt.Printf("Error dialing L2 node: %v\n", err)
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
 	config := utils.BatchDecoderConfig{
 		L2ChainID:    new(big.Int).SetUint64(req.L2ChainID),
-		L2Node:       req.L2Node,
-		L1RPC:        req.L1RPC,
-		L1Beacon:     req.L1Beacon,
+		L2Node:       l2Node,
+		L1RPC:        *l1Client,
+		L1Beacon:     l1BeaconClient,
 		BatchSender:  common.HexToAddress(req.BatchSender),
 		L2StartBlock: req.StartBlock,
 		L2EndBlock:   req.EndBlock,
-		DataDir: fmt.Sprintf("/tmp/batch_decoder/%d/transactions_cache", req.L2ChainID),
+		DataDir:      fmt.Sprintf("/tmp/batch_decoder/%d/transactions_cache", req.L2ChainID),
 	}
 
-	ranges, err := utils.GetAllSpanBatchesInBlockRange(config)
+	ranges, err := utils.GetAllSpanBatchesInL2BlockRange(config)
 	if err != nil {
 		fmt.Printf("Error getting span batch ranges: %v\n", err)
 		http.Error(w, err.Error(), http.StatusInternalServerError)
