+++ op-succinct-proposer/Cargo.toml
@@ -27,6 +27,10 @@ path = "bin/cost_estimator.rs"
 name = "load_stdin"
 path = "bin/load_stdin.rs"
 
+[[bin]]
+name = "span_batch_ranges"
+path = "bin/span_batch_ranges.rs"
+
 [dependencies]
 
 # workspace
