+++ op-succinct-proposer/build.rs
@@ -68,6 +68,7 @@ fn main() {
     let programs = vec!["range"];
 
     for program in programs {
+        // Note: Don't comment this out, because the Docker program depends on the native program for range being built.
         build_native_program(program);
         // build_zkvm_program(program);
     }
