+++ zkvm-host/build.rs
@@ -39,11 +39,11 @@ fn build_zkvm_program(program: &str) {
 }
 
 fn main() {
-    // let programs = vec!["fault-proof", "range"];
-    // for program in programs {
-    //     build_native_program(program);
-    //     build_zkvm_program(program);
-    // }
+    let programs = vec!["range"];
+    for program in programs {
+        // build_native_program(program);
+        // build_zkvm_program(program);
+    }
 
     // build_zkvm_program("aggregation");
 }
